Initial Review backend
I made the first iteration of the review system. Currently new reviews can only be added in the admin page but making a new page for that should not be too difficult. All of the reviews are stored in a sqlite database and have a foreign key to a user listed as the author and a posted date that automatically gets filled in when the post gets created.